--- a/common/herrors/errors.go
+++ b/common/herrors/errors.go
@@ -57,6 +57,11 @@ func PrintStackTrace(w io.Writer) {
 	fmt.Fprintf(w, "%s", buf)
 }
 
+// ErrorSender is a, typically, non-blocking error handler.
+type ErrorSender interface {
+	SendError(err error)
+}
+
 // Recover is a helper function that can be used to capture panics.
 // Put this at the top of a method/function that crashes in a template:
 //     defer herrors.Recover()