--- a/common/para/para_test.go
+++ b/common/para/para_test.go
@@ -16,6 +16,7 @@ package para
 import (
 	"context"
 	"runtime"
+
 	"sort"
 	"sync"
 	"sync/atomic"