--- a/hugolib/page__paths.go
+++ b/hugolib/page__paths.go
@@ -34,14 +34,10 @@ func newPagePaths(
 
 	outputFormats := pm.outputFormats()
 	if len(outputFormats) == 0 {
-		outputFormats = pm.s.outputFormats[pm.Kind()]
-	}
-
-	if len(outputFormats) == 0 {
 		return pagePaths{}, nil
 	}
 
-	if pm.headless {
+	if pm.noRender() {
 		outputFormats = outputFormats[:1]
 	}
 
@@ -55,9 +51,9 @@ func newPagePaths(
 
 		var relPermalink, permalink string
 
-		// If a page is headless or bundled in another, it will not get published
-		// on its own and it will have no links.
-		if !pm.headless && !pm.bundled {
+		// If a page is headless or marked as "no render", or bundled in another,
+		// it will not get published on its own and it will have no links.
+		if !pm.noRender() && !pm.bundled {
 			relPermalink = paths.RelPermalink(s.PathSpec)
 			permalink = paths.PermalinkForOutputFormat(s.PathSpec, f)
 		}
@@ -77,8 +73,14 @@ func newPagePaths(
 
 	}
 
+	var out page.OutputFormats
+	if !pm.noRender() {
+		out = pageOutputFormats
+	}
+
 	return pagePaths{
-		outputFormats:        pageOutputFormats,
+		outputFormats:        out,
+		firstOutputFormat:    pageOutputFormats[0],
 		targetPaths:          targets,
 		targetPathDescriptor: targetPathDescriptor,
 	}, nil
@@ -86,7 +88,8 @@ func newPagePaths(
 }
 
 type pagePaths struct {
-	outputFormats page.OutputFormats
+	outputFormats     page.OutputFormats
+	firstOutputFormat page.OutputFormat
 
 	targetPaths          map[string]targetPathsHolder
 	targetPathDescriptor page.TargetPathDescriptor